restructured landing page
