+++ src/components/React-Slick/react-slick.scss
@@ -1,0 +1,36 @@
+@import "../../colors.scss";
+@import "../../mixins.scss";
+.home-modal {
+  width: 100vw;
+  height: 100vh;
+  background: rgba(0, 0, 0, 0.879);
+  position: absolute;
+  top: 0;
+  left: 0;
+  display: flex;
+  flex-flow: column nowrap;
+  justify-content: space-evenly;
+  align-items: center;
+  z-index: 1000;
+  button {
+    border: none;
+    width: 50px;
+    height: 50px;
+    border-radius: 50%;
+    font-size: 2rem;
+    background: $grey;
+    cursor: pointer;
+    transition: transform 1s;
+  }
+  button:hover {
+    transform: scale(1.1);
+  }
+  button:active {
+    transform: scale(0.5);
+  }
+  .show-image {
+    width: 80%;
+    height: 50%;
+    object-fit: contain;
+  }
+}
